WIP: save current state before rebase
