+++ .idea/libraries/Flutter_Plugins.xml
@@ -1,39 +1,41 @@
 <component name="libraryTable">
   <library name="Flutter Plugins" type="FlutterPluginsLibraryType">
     <CLASSES>
-      <root url="file://$PROJECT_DIR$/../.pub-cache/hosted/pub.dev/permission_handler-12.0.1" />
-      <root url="file://$PROJECT_DIR$/../.pub-cache/hosted/pub.dev/sqflite_darwin-2.4.2" />
-      <root url="file://$PROJECT_DIR$/../.pub-cache/hosted/pub.dev/google_sign_in-6.3.0" />
-      <root url="file://$PROJECT_DIR$/../.pub-cache/hosted/pub.dev/google_sign_in_ios-5.9.0" />
-      <root url="file://$PROJECT_DIR$/../.pub-cache/hosted/pub.dev/sqflite-2.4.2" />
-      <root url="file://$PROJECT_DIR$/../.pub-cache/hosted/pub.dev/path_provider_linux-2.2.1" />
-      <root url="file://$PROJECT_DIR$/../.pub-cache/hosted/pub.dev/flutter_secure_storage_web-1.2.1" />
-      <root url="file://$PROJECT_DIR$/../.pub-cache/hosted/pub.dev/path_provider_windows-2.3.0" />
-      <root url="file://$PROJECT_DIR$/../.pub-cache/hosted/pub.dev/path_provider-2.1.5" />
-      <root url="file://$PROJECT_DIR$/../.pub-cache/hosted/pub.dev/permission_handler_html-0.1.3+5" />
-      <root url="file://$PROJECT_DIR$/../.pub-cache/hosted/pub.dev/google_sign_in_android-6.2.1" />
-      <root url="file://$PROJECT_DIR$/../.pub-cache/hosted/pub.dev/google_sign_in_web-0.12.4+4" />
-      <root url="file://$PROJECT_DIR$/../.pub-cache/hosted/pub.dev/flutter_secure_storage-9.2.4" />
-      <root url="file://$PROJECT_DIR$/../.pub-cache/hosted/pub.dev/flutter_secure_storage_windows-3.1.2" />
-      <root url="file://$PROJECT_DIR$/../.pub-cache/hosted/pub.dev/shared_preferences_windows-2.4.1" />
-      <root url="file://$PROJECT_DIR$/../.pub-cache/hosted/pub.dev/sqflite_android-2.4.2+3" />
-      <root url="file://$PROJECT_DIR$/../.pub-cache/hosted/pub.dev/shared_preferences_linux-2.4.1" />
-      <root url="file://$PROJECT_DIR$/../.pub-cache/hosted/pub.dev/shared_preferences_web-2.4.3" />
-      <root url="file://$PROJECT_DIR$/../.pub-cache/hosted/pub.dev/image_picker_for_web-3.1.1" />
-      <root url="file://$PROJECT_DIR$/../.pub-cache/hosted/pub.dev/shared_preferences-2.5.5" />
-      <root url="file://$PROJECT_DIR$/../.pub-cache/hosted/pub.dev/file_selector_macos-0.9.5" />
-      <root url="file://$PROJECT_DIR$/../.pub-cache/hosted/pub.dev/image_picker_windows-0.2.2" />
-      <root url="file://$PROJECT_DIR$/../.pub-cache/hosted/pub.dev/path_provider_foundation-2.5.1" />
-      <root url="file://$PROJECT_DIR$/../.pub-cache/hosted/pub.dev/shared_preferences_foundation-2.5.6" />
-      <root url="file://$PROJECT_DIR$/../.pub-cache/hosted/pub.dev/image_picker-1.2.1" />
-      <root url="file://$PROJECT_DIR$/../.pub-cache/hosted/pub.dev/image_picker_ios-0.8.13+6" />
-      <root url="file://$PROJECT_DIR$/../.pub-cache/hosted/pub.dev/image_picker_linux-0.2.2" />
-      <root url="file://$PROJECT_DIR$/../.pub-cache/hosted/pub.dev/path_provider_android-2.2.23" />
-      <root url="file://$PROJECT_DIR$/../.pub-cache/hosted/pub.dev/file_selector_windows-0.9.3+5" />
-      <root url="file://$PROJECT_DIR$/../.pub-cache/hosted/pub.dev/image_picker_macos-0.2.2+1" />
-      <root url="file://$PROJECT_DIR$/../.pub-cache/hosted/pub.dev/shared_preferences_android-2.4.23" />
-      <root url="file://$PROJECT_DIR$/../.pub-cache/hosted/pub.dev/file_selector_linux-0.9.4" />
-      <root url="file://$PROJECT_DIR$/../.pub-cache/hosted/pub.dev/image_picker_android-0.8.13+15" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/image_picker_linux-0.2.2" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/path_provider_android-2.3.1" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/image_picker_for_web-3.1.1" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/shared_preferences-2.5.5" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/file_selector_linux-0.9.4" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/path_provider_linux-2.2.1" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/path_provider-2.1.5" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/path_provider_windows-2.3.0" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/image_picker_windows-0.2.2" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/image_picker_macos-0.2.2+1" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/image_picker_ios-0.8.13+6" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/google_sign_in-6.3.0" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/google_sign_in_ios-5.9.0" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/jni_flutter-1.0.1" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/sqflite_android-2.4.2+3" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/shared_preferences_foundation-2.5.6" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/flutter_secure_storage_windows-3.1.2" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/permission_handler-12.0.1" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/sqflite-2.4.2" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/flutter_secure_storage_web-1.2.1" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/shared_preferences_web-2.4.3" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/permission_handler_html-0.1.3+5" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/flutter_secure_storage-9.2.4" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/image_picker_android-0.8.13+16" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/file_selector_macos-0.9.5" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/shared_preferences_windows-2.4.1" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/google_sign_in_web-0.12.4+4" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/shared_preferences_android-2.4.23" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/file_selector_windows-0.9.3+5" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/path_provider_foundation-2.6.0" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/image_picker-1.2.1" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/shared_preferences_linux-2.4.1" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/sqflite_darwin-2.4.2" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/google_sign_in_android-6.2.1" />
+      <root url="file://$USER_HOME$/AppData/Local/Pub/Cache/hosted/pub.dev/jni-1.0.0" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
